Rubber cloud providers can now optionally define Factory classes.  Did this for Aws and reorganized AWS Classic and VPC implementations
