+++ lib/rubber/cloud.rb
@@ -7,15 +7,17 @@ module Rubber
       require "rubber/cloud/#{provider}"
       provider_env = env.cloud_providers[provider]
 
-      if provider_env.vpc
-        require "rubber/cloud/aws_vpc"
-        clazz = Rubber::Cloud::AwsVpc
-      else
+      # Check to see if we have a Rubber::Cloud::Provider::Factory class.  If
+      # not, fall back to Rubber::Cloud::Provider
+      begin
+        factory = Rubber::Cloud.const_get(Rubber::Util.camelcase(provider))::Factory
+        return factory.get_provider(provider_env, capistrano)
+      rescue NameError => e
         clazz = Rubber::Cloud.const_get(Rubber::Util.camelcase(provider))
+        return clazz.new(provider_env, capistrano)
       end
-
-      return clazz.new(provider_env, capistrano)
     end
 
   end
 end
+
